requirements.txt
flask
beautifulsoup4
pymongo
python-dotenv
playwright